requirements.txt
azure-core==1.27.0
azure-identity==1.13.0
azure-storage-blob==12.17.0b1
azure-storage-file-datalake==12.11.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
configparser==5.3.0
confluent-kafka==2.1.1
cryptography==41.0.1
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.3
